refactor(artifact): remove unused `create` package and associated tests
- Deleted `pkg/artifact/create.go` and `pkg/artifact/create_test.go` as they are unused and redundant.

+++ internal/cli/helpers.go
@@ -65,7 +65,7 @@ func loadIndexManager(config *config.Config, httpClient http.Client) index.Manag
 		repositories = append(repositories, &index.Repository{
 			Name:     repo.Name,
 			URL:      repo.GetURL(),
-			Priority: uint(repo.Priority),
+			Priority: repo.Priority,
 			Enabled:  repo.Enabled,
 		})
 	}
